Split stats into modules.
